PR #5 feedback: centralize access to localStorage into src/store/index.ts
